functions/func-process/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-eventhub-checkpointstoreblob-aio
python-dotenv==1.0.1
azure-storage-file-datalake==12.14.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-vision-imageanalysis==1.0.0b2
bs4==0.0.1